+++ lib/castle/core/process_response.rb
@@ -9,9 +9,10 @@ module Castle
         401 => Castle::UnauthorizedError,
         403 => Castle::ForbiddenError,
         404 => Castle::NotFoundError,
-        419 => Castle::UserUnauthorizedError,
-        422 => Castle::InvalidParametersError
+        419 => Castle::UserUnauthorizedError
       }.freeze
+
+      INVALID_REQUEST_TOKEN = 'invalid_request_token'
 
       class << self
         # @param response [Response]
@@ -36,8 +37,29 @@ module Castle
 
           raise Castle::InternalServerError if response.code.to_i.between?(500, 599)
 
+          raise_error422(response) if response.code.to_i == 422
+
           error = RESPONSE_ERRORS.fetch(response.code.to_i, Castle::ApiError)
-          raise error, response[:message]
+
+          raise error
+        end
+
+        def raise_error422(response)
+          if response.body
+            begin
+              parsed_body = JSON.parse(response.body, symbolize_names: true)
+              if parsed_body.is_a?(Hash) && parsed_body.key?(:type)
+                if parsed_body[:type] == INVALID_REQUEST_TOKEN
+                  raise Castle::InvalidRequestTokenError, parsed_body[:message]
+                else
+                  raise Castle::InvalidParametersError, parsed_body[:message]
+                end
+              end
+            rescue JSON::ParserError
+            end
+          end
+
+          raise Castle::InvalidParametersError
         end
       end
     end