+++ lib/castle/configuration.rb
@@ -45,7 +45,6 @@ module Castle
       Te
       Upgrade-Insecure-Requests
       User-Agent
-      X-Castle-Client-Id
       X-Requested-With
     ].freeze